Refactor out into seperate files.
Rename methods from `handle` to `on`
