+++ package.json
@@ -10,6 +10,11 @@
   "author": "",
   "license": "MIT",
   "devDependencies": {
+    "babel-core": "^6.13.2",
+    "babel-loader": "^6.2.4",
+    "babel-preset-es2015": "^6.13.2",
+    "babel-preset-react": "^6.11.1",
+    "babel-preset-stage-1": "^6.13.0",
     "http-server": "^0.9.0",
     "webpack": "^1.13.1"
   },
