+++ src/toolbar.js
@@ -1,0 +1,53 @@
+import React, { Component } from 'react'
+
+class Toolbar extends Component {
+
+  constructor(props) {
+
+    super(props)
+
+    this.state = {
+      links: props.links,
+      onLinkClick: props.onLinkClick
+    }
+
+    this.onLinkClick = this.onLinkClick.bind(this)
+
+  }
+
+  static propTypes() {
+    return {
+      onLinkClick: React.PropTypes.func.isRequired,
+      links: React.PropTypes.array.isRequired,
+    }
+  }
+
+  onLinkClick(evt) {
+    this.state.onLinkClick(evt.target.innerText)
+  }
+
+  render() {
+    return React.createElement(
+      'header',
+      {},
+      React.createElement(
+        'ul',
+        {},
+        this.state.links.map((item, index) => {
+          return React.createElement(
+            'li',
+            { key: index },
+            React.createElement(
+              'a',
+              { onClick: this.onLinkClick },
+              item
+            )
+          )
+        })
+      )
+    )
+  }
+
+}
+
+export default Toolbar
